Split V9 public types into fact and internal modules
Move public fact primitives into safety-score-v9-public-facts.ts and the card cross-field refinements into safety-score-v9-public-internal.ts, retaining every root export so all importers compile unchanged; the root drops from 2283 to about 1720 lines. Byte-identical moved code verified; 153 pinned tests pass with zero fixture changes. Audit findings R-13/S-11 (P3-04 tranche 4).
